✨ (Enhance): AuthButtons 및 IntroSlider 컴포넌트 추가 및 SignInPage에서 통합
